+++ Changelog.md
@@ -6,6 +6,9 @@ and as of v1.0.0 this project adheres to [Semantic Versioning](https://semver.or
 
 ## [Unreleased]
 
+### Added
+- Support and testing for Rails 7.2 & Rails main.
+
 ### Removed
 - Support for ActiveRecord's legacy connection handling.